+++ src/routes/team.routes.js
@@ -3,7 +3,7 @@ const { applyRouteValidation } = require('../middleware/requestValidation.middle
 const routeSchemas = require('../schemas/team.routes.schema');
 const { authorizeFirmPermission } = require('../middleware/permission.middleware');
 const { requirePrimaryAdmin } = require('../middleware/rbac.middleware');
-const { listTeams, createTeam, updateTeam, assignUserToTeam } = require('../controllers/team.controller');
+const { listTeams, createTeam, updateTeam, assignUserToTeam, addUserToQcWorkbasket } = require('../controllers/team.controller');
 
 const router = applyRouteValidation(express.Router(), routeSchemas);
 
@@ -11,5 +11,6 @@ router.get('/', authorizeFirmPermission('CASE_VIEW'), listTeams);
 router.post('/', requirePrimaryAdmin, createTeam);
 router.patch('/:id', requirePrimaryAdmin, updateTeam);
 router.post('/:id/assign-user', requirePrimaryAdmin, assignUserToTeam);
+router.post('/:id/qc/add-user', authorizeFirmPermission('USER_MANAGE'), addUserToQcWorkbasket);
 
 module.exports = router;